+++ tests/servlet_modules/application/servlets/reload.jss.mjs
@@ -1,0 +1,3 @@
+export function servlet() {
+    return "reload one";
+}